Implemented some game logic
Implemented all the platform effects
Implemented speed-variable fov
Implemented score system
Increased lane spacing
Added game state variable
Solved a few warnings in the code

+++ Source/src/Shaders/Base.FS.glsl
@@ -10,7 +10,7 @@ uniform vec3 eye_position;
 
 uniform float material_kd;
 uniform float material_ks;
-uniform int material_shininess;
+uniform float material_shininess;
 
 uniform vec3 object_color;
